chore: Entity 설정 수정

See merge request s10-ai-image-sub2/S10P22A302!5

+++ BackEnd/IdealMaker/src/main/java/com/ideal/idealmaker/ideal/domain/IdealAnimalTypeId.java
@@ -21,5 +21,5 @@ public class IdealAnimalTypeId implements Serializable {
 	private Long idealId;
 
 	@Column(name = "animal_type_id")
-	private Long animalTypeId;
+	private Integer animalTypeId;
 }
